requirements.txt
torch==2.6.0 # SHOULD BE DONE OUTSIDE OF REQUIREMENTS.TXT
accelerate==0.31.0
datasets==3.2.0
deepspeed==0.16.3
flash-attn==2.6.3
jsonargparse==4.36.0
openai==1.60.0
protobuf==5.29.3
python-dotenv==1.0.1
rich==13.9.4
sentencepiece==0.2.0
transformers==4.48.3
matplotlib==3.10.1